refactor: clarify peer management ownership (#1472)
Part of #1472
