+++ internal/peermanagement/overlay.go
@@ -476,116 +476,6 @@ func (o *Overlay) NotePeerHasTxSet(peerID PeerID, hash [32]byte) bool {
 }
 
 // SetLedgerHintProvider wires the hint source; nil suppresses headers.
-func (o *Overlay) SetLedgerHintProvider(fn func() (LedgerHints, bool)) {
-	o.providersMu.Lock()
-	o.ledgerHintProvider = fn
-	o.providersMu.Unlock()
-}
-
-func (o *Overlay) ledgerHintProviderSnapshot() func() (LedgerHints, bool) {
-	o.providersMu.RLock()
-	defer o.providersMu.RUnlock()
-	return o.ledgerHintProvider
-}
-
-// SetValidLedgerProvider wires the validated-ledger source used by
-// handleStatusChange. ok=false suppresses tracking updates.
-func (o *Overlay) SetValidLedgerProvider(fn func() (seq uint32, age time.Duration, ok bool)) {
-	o.providersMu.Lock()
-	o.validLedgerProvider = fn
-	o.providersMu.Unlock()
-}
-
-func (o *Overlay) validLedgerProviderSnapshot() func() (seq uint32, age time.Duration, ok bool) {
-	o.providersMu.RLock()
-	defer o.providersMu.RUnlock()
-	return o.validLedgerProvider
-}
-
-// PeerStatusUpdate captures the post-decode TMStatusChange fields the
-// RPC layer needs to materialize a peer_status WebSocket event. Pointer
-// fields preserve protobuf has-presence; nil means the wire field was
-// absent and the RPC layer omits the JSON field.
-type PeerStatusUpdate struct {
-	// Status is the UPPERCASE status name. Carries the
-	// post-inheritance value returned by applyStatusChange, so a
-	// status-less wire message still emits the prior enum once.
-	Status string
-	// Action is CLOSING_LEDGER, ACCEPTED_LEDGER or SWITCHED_LEDGER.
-	// LOST_SYNC is unreachable because handleStatusChange returns
-	// before the publish.
-	Action string
-	// LedgerHash is sourced from the peer's post-apply closed-ledger
-	// state rather than echoing the raw wire bytes. When the wire
-	// bytes were malformed that state is cleared and the 64-char zero
-	// hex string is emitted — so callers must ALWAYS emit a value
-	// when the wire carried the field, falling back to "00…00".
-	LedgerHash string
-	// LedgerIndex: nil = field absent; non-nil = emit (even when
-	// value is 0 — a peer can legitimately advertise the genesis seq).
-	LedgerIndex *uint32
-	// Date is auto-stamped with the local clock when the wire didn't
-	// carry a networktime, so it is always non-nil here.
-	Date *uint32
-	// LedgerIndexMin / LedgerIndexMax are nil unless both wire fields
-	// were present.
-	LedgerIndexMin *uint32
-	LedgerIndexMax *uint32
-}
-
-// SetPeerStatusPublisher wires a sink for peer_status events. The
-// overlay invokes this callback for every non-lostSync TMStatusChange
-// after state has been recorded. Passing nil disconnects the sink.
-func (o *Overlay) SetPeerStatusPublisher(fn func(PeerStatusUpdate)) {
-	o.providersMu.Lock()
-	o.peerStatusPublisher = fn
-	o.providersMu.Unlock()
-}
-
-func (o *Overlay) peerStatusPublisherSnapshot() func(PeerStatusUpdate) {
-	o.providersMu.RLock()
-	defer o.providersMu.RUnlock()
-	return o.peerStatusPublisher
-}
-
-// peerStatusUpperName returns the UPPERCASE status name
-// (CONNECTING/...) emitted by peer_status events, distinct from the
-// lowercase strings used by the `peers` RPC. Returns "" for nsUNKNOWN
-// or any unknown enum.
-func peerStatusUpperName(s message.NodeStatus) string {
-	switch s {
-	case message.NodeStatusConnecting:
-		return "CONNECTING"
-	case message.NodeStatusConnected:
-		return "CONNECTED"
-	case message.NodeStatusMonitoring:
-		return "MONITORING"
-	case message.NodeStatusValidating:
-		return "VALIDATING"
-	case message.NodeStatusShutting:
-		return "SHUTTING"
-	default:
-		return ""
-	}
-}
-
-// peerStatusActionName maps a NodeEvent to its peer_status action
-// name. handleStatusChange returns before the publish for neLOST_SYNC,
-// so the LOST_SYNC arm is unreachable from this call site and
-// intentionally omitted. Unknown enums fall through silently.
-func peerStatusActionName(e message.NodeEvent) string {
-	switch e {
-	case message.NodeEventClosingLedger:
-		return "CLOSING_LEDGER"
-	case message.NodeEventAcceptedLedger:
-		return "ACCEPTED_LEDGER"
-	case message.NodeEventSwitchedLedger:
-		return "SWITCHED_LEDGER"
-	default:
-		return ""
-	}
-}
-
 // generateInstanceCookie draws a cookie uniform in [1, MaxUint64];
 // only 0 is rejected.
 func generateInstanceCookie() (uint64, error) {
